refactor: 적절한 클래스로 테스트 코드 이동
