+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/annotation/AnnotationHandlerMappingTest.java
@@ -6,8 +6,11 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
+import com.interface21.webmvc.servlet.mvc.tobe.HandlerKey;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.lang.reflect.Field;
+import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -22,6 +25,17 @@ class AnnotationHandlerMappingTest {
     void setUp() {
         handlerMapping = new AnnotationHandlerMapping("samples");
         handlerMapping.initialize();
+    }
+
+    @Test
+    @DisplayName("매핑은 HTTP 요청을 처리할 핸들러 목록을 초기화한다.")
+    void getHandler() throws NoSuchFieldException, IllegalAccessException {
+        Field filed = handlerMapping.getClass().getDeclaredField("handlerExecutions");
+        filed.setAccessible(true);
+
+        final var handlerExecutions = (Map<HandlerKey, HandlerExecution>) filed.get(handlerMapping);
+
+        assertThat(handlerExecutions.keySet()).hasSize(12);
     }
 
     @Test
